DropDownEditors width option refactoring
