+++ js/ui/date_box/ui.date_box.base.js
@@ -259,7 +259,13 @@ const DateBox = DropDownEditor.inherit({
     _renderDimensions: function() {
         this.callBase();
         this.$element().toggleClass(DX_AUTO_WIDTH_CLASS, !this.option('width'));
-        this._strategy?._dimensionChanged();
+        this._dimensionChanged();
+    },
+
+    _dimensionChanged: function() {
+        this.callBase(arguments);
+
+        this._popup && this._strategy._updatePopupHeight?.();
     },
 
     _refreshFormatClass: function() {
@@ -354,7 +360,7 @@ const DateBox = DropDownEditor.inherit({
         const popupConfig = this.callBase();
         return extend(this._strategy.popupConfig(popupConfig), {
             title: this._getPopupTitle(),
-            dragEnabled: false
+            dragEnabled: false,
         });
     },
