+++ js/ui/date_box/ui.date_box.strategy.calendar_with_time.js
@@ -138,6 +138,7 @@ const CalendarWithTimeStrategy = CalendarStrategy.inherit({
     popupConfig: function(popupConfig) {
         const calendarPopupConfig = this.callBase(popupConfig);
         const result = extend(calendarPopupConfig, {
+            width: 'auto',
             onShowing: (function() {
                 if(this._box.option('_layoutStrategy') === 'fallback') {
                     const clockMinWidth = this._getPopup().$content().find('.dx-timeview-clock').css('minWidth');
