+++ js/ui/date_box/ui.date_box.strategy.date_view.js
@@ -28,6 +28,7 @@ const DateViewStrategy = DateBoxStrategy.inherit({
         return {
             toolbarItems: this.dateBox._popupToolbarItemsConfig(),
             onInitialized: config.onInitialized,
+            width: 'auto',
 
             defaultOptionsRules: [
                 {
