+++ js/ui/drop_down_box.js
@@ -178,10 +178,6 @@ const DropDownBox = DropDownEditor.inherit({
         return deferred.promise();
     },
 
-    _updatePopupWidth: function() {
-        this._setPopupOption('width', this.$element().outerWidth());
-    },
-
     _popupElementTabHandler: function(e) {
         if(normalizeKeyName(e) !== 'tab') return;
 
@@ -272,9 +268,6 @@ const DropDownBox = DropDownEditor.inherit({
         const { focusStateEnabled } = this.option();
 
         return extend(this.callBase(), {
-            width: function() {
-                return this.$element().outerWidth();
-            }.bind(this),
             height: 'auto',
             tabIndex: -1,
             dragEnabled: false,
@@ -304,10 +297,6 @@ const DropDownBox = DropDownEditor.inherit({
     _optionChanged: function(args) {
         this._dataExpressionOptionChanged(args);
         switch(args.name) {
-            case 'width':
-                this.callBase(args);
-                this._popup && this._popup.repaint();
-                break;
             case 'dataSource':
                 this._renderInputValue();
                 break;
