+++ js/ui/drop_down_editor/ui.drop_down_editor.js
@@ -392,6 +392,10 @@ const DropDownEditor = TextBox.inherit({
         }
     },
 
+    _getInputWidth: function() {
+        return this.$element().outerWidth();
+    },
+
     _attachFocusOutHandler: function() {
         if(isIOs) {
             this._detachFocusOutEvents();
@@ -519,7 +523,7 @@ const DropDownEditor = TextBox.inherit({
                 of: this.$element()
             }),
             showTitle: this.option('dropDownOptions.showTitle'),
-            width: 'auto',
+            width: this._getInputWidth.bind(this),
             height: 'auto',
             shading: false,
             closeOnTargetScroll: true,
@@ -538,6 +542,20 @@ const DropDownEditor = TextBox.inherit({
         };
     },
 
+
+    _getPopupWidth() {
+        const popupWidth = this.option('dropDownOptions.width');
+
+        if(popupWidth === null) {
+            return undefined;
+        }
+        if(typeof popupWidth === 'function') {
+            return popupWidth();
+        }
+
+        return popupWidth;
+    },
+
     _popupInitializedHandler: function() {
         if(!this.option('onPopupInitialized')) {
             return;
@@ -546,6 +564,14 @@ const DropDownEditor = TextBox.inherit({
         return (e) => {
             this._popupInitializedAction({ popup: e.component });
         };
+    },
+
+    _dimensionChanged: function() {
+        const popupWidth = this._getPopupWidth();
+
+        if(popupWidth === undefined) {
+            this._setPopupOption('width', this._getInputWidth.bind(this));
+        }
     },
 
     _popupPositionedHandler: function(e) {
@@ -704,15 +730,14 @@ const DropDownEditor = TextBox.inherit({
         this.option('focusStateEnabled') && this.focus();
     },
 
-    _updatePopupWidth: noop,
-
     _popupOptionChanged: function(args) {
         const options = Widget.getOptionsFromContainer(args);
 
         this._setPopupOption(options);
 
-        if(Object.keys(options).indexOf('width') !== -1 && options['width'] === undefined) {
-            this._updatePopupWidth();
+        const optionsKeys = Object.keys(options);
+        if(optionsKeys.indexOf('width') !== -1 || optionsKeys.indexOf('height') !== -1) {
+            this._dimensionChanged();
         }
     },
 
@@ -751,6 +776,10 @@ const DropDownEditor = TextBox.inherit({
 
     _optionChanged: function(args) {
         switch(args.name) {
+            case 'width':
+                this.callBase(args);
+                this._popup?.repaint();
+                break;
             case 'opened':
                 this._renderOpenedState();
                 break;
